+++ src/backend/routes/audio_routes.py
@@ -2,7 +2,7 @@
 import logging
 import requests
 import json
-from flask import Blueprint, request, jsonify, g, Response
+from flask import Blueprint, request, jsonify, g, Response,session
 
 from backend.services.audioService import AudioService
 from backend.services.subscriptionService import SubscriptionService
@@ -91,6 +91,51 @@ def audio_analysis():
 
     except Exception as e:
         logger.error(f"Error analyzing audio: {str(e)}")
+        return jsonify({"error": str(e)}), 500
+
+@audio_bp.route("/voice_isolate", methods=["POST"])
+def isolate_voice():
+    if "audio" not in request.files or "episode_id" not in request.form:
+        return jsonify({"error": "Audio file and episode_id are required"}), 400
+
+    user_id = session.get("user_id")
+    if not user_id:
+        return jsonify({"error": "Authentication required"}), 401
+
+    # Consume credits BEFORE processing
+    try:
+        consume_credits(user_id, "voice_isolation")
+    except ValueError as e:
+        logger.warning(f"User {user_id} has insufficient credits for voice isolation.")
+        return jsonify({
+            "error": str(e),
+            "redirect": "/store"
+        }), 403
+
+    audio_file = request.files["audio"]
+    episode_id = request.form["episode_id"]
+    filename = audio_file.filename
+    audio_bytes = audio_file.read()
+
+    try:
+        blob_url = audio_service.isolate_voice(audio_bytes, filename, episode_id)
+        return jsonify({"isolated_blob_url": blob_url})  
+    except Exception as e:
+        logger.error(f"Error during voice isolation: {str(e)}")
+        return jsonify({"error": str(e)}), 500
+
+    
+@audio_bp.route("/get_isolated_audio", methods=["GET"])
+def get_isolated_audio():
+    url = request.args.get("url")
+    if not url:
+        return jsonify({"error": "Missing URL"}), 400
+    
+    try:
+        response = requests.get(url)
+        return Response(response.content, content_type="audio/wav")
+    
+    except Exception as e:
         return jsonify({"error": str(e)}), 500
